Struct kubernix::Config

pub struct Config { /* fields omitted */ }

The global configuration

Methods

impl Config

pub fn root(&self) -> &Path

Getter for root

pub fn subcommand(&self) -> &Option<SubCommand>

Getter for subcommand

pub fn log_level(&self) -> LevelFilter

Getter for log_level

pub fn nodes(&self) -> u8

Getter for nodes

pub fn overlay(&self) -> &Option<PathBuf>

Getter for overlay

pub fn packages(&self) -> &[String]

Getter for packages

pub fn cidr(&self) -> Ipv4Network

Getter for cidr

pub fn container_runtime(&self) -> &str

Getter for container_runtime

pub fn no_shell(&self) -> bool

Getter for no_shell

pub fn canonicalize_root(&mut self) -> Result<()>

Make the configs root path absolute

pub fn to_file(&self) -> Result<()>

Write the current configuration to the internal set root path

pub fn try_load_file(&mut self) -> Result<()>

Read the configuration from the internal set root path If not existing, write the current configuration to the path.

pub fn shell_ok(&self) -> Result<String>

Return the set shell as result type

pub fn multi_node(&self) -> bool

Returns true if multi node support is enabled