Struct kubernix::Config

pub struct Config { /* fields omitted */ }

The global configuration

Implementations

impl Config

pub fn log_level(&self) -> LevelFilter

The logging level of the application

pub fn cidr(&self) -> Ipv4Network

The CIDR used for the cluster

pub fn nodes(&self) -> u8

The number of nodes to be registered

impl Config

pub fn subcommand(&self) -> &Option<SubCommand>

All available subcommands

pub fn root(&self) -> &PathBuf

Path where all the runtime data is stored

pub fn overlay(&self) -> &Option<PathBuf>

The Nix package overlay to be used

pub fn packages(&self) -> &Vec<String>

Additional dependencies to be added to the environment

pub fn shell(&self) -> &Option<String>

The shell executable to be used, defaults to $SHELL, fallback is sh

pub fn container_runtime(&self) -> &String

The container runtime to be used for the nodes, irrelevant if nodes equals to 1

pub fn no_shell(&self) -> &bool

Do not spawn an interactive shell after bootstrap

impl Config

pub fn canonicalize_root(&mut self) -> Result<()>

Make the configs root path absolute

pub fn to_file(&self) -> Result<()>

Write the current configuration to the internal set root path

pub fn try_load_file(&mut self) -> Result<()>

Read the configuration from the internal set root path If not existing, write the current configuration to the path.

pub fn shell_ok(&self) -> Result<String>

Return the set shell as result type

pub fn multi_node(&self) -> bool

Returns true if multi node support is enabled